% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{assoc_tibble.data.frame}
\alias{assoc_tibble.data.frame}
\title{A tibble structure for a measure of association}
\usage{
\method{assoc_tibble}{data.frame}(data, measure_type = NA_character_)
}
\arguments{
\item{data}{dataframe A dataset for exploring association among the variables or a dataframe with calculated
measure of association for every variable pair}

\item{measure_type}{a character string indicating the measure of association}
}
\value{
tibble
}
\description{
Creates a tibble for every variable pair in a dataset with a measure of association
}
