% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{assoc_tibble}
\alias{assoc_tibble}
\alias{assoc_tibble.matrix}
\alias{assoc_tibble.data.frame}
\title{A generic function to create a data structure for every variable pair in a dataset}
\usage{
assoc_tibble(data, measure_type = "?", pair_type = "?")

\method{assoc_tibble}{matrix}(data, measure_type = "?", pair_type = "?")

\method{assoc_tibble}{data.frame}(data, measure_type = NA_character_, pair_type = NA_character_)
}
\arguments{
\item{data}{A dataframe.}

\item{measure_type}{a character string indicating the measure of association.}

\item{pair_type}{a character string specifying the type of variable pair.}
}
\value{
A data structure for pairs of variables with a column \code{measure} for measure value,
\code{measure_type} for a type of association measure and \code{pair_type} for the variable pair.
}
\description{
Creates a data structure for every variable pair in a dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{assoc_tibble(matrix)}: assoc_tibble method

\item \code{assoc_tibble(data.frame)}: assoc_tibble method

}}
\examples{
assoc_tibble(cor(iris[,1:4]), measure_type="pearson")
assoc_tibble(iris)
}
