% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_heatmap.R
\name{association_heatmap}
\alias{association_heatmap}
\title{An association matrix plot (similar to popular correlation matrix plot)}
\usage{
association_heatmap(
  lassoc,
  uassoc = NULL,
  var_order = "default",
  limits = c(-1, 1)
)
}
\arguments{
\item{lassoc}{A tibble with the calculated association measures for the lower triangle of the matrix plot.}

\item{uassoc}{A tibble with the calculated association measures for the upper triangle of the matrix plot.
If \emph{NULL} (default) the matrix plot is symmetric.}

\item{var_order}{a character string for the ordering of the variables. Either "default" (default) or NA.}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a
conventional way.
}
\examples{
association_heatmap(calc_assoc(iris))
}
