% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_measure.R
\name{association_measure}
\alias{association_measure}
\title{Find correlation data structure}
\usage{
association_measure(
  data,
  group_by = NULL,
  use = "everything",
  method = "pearson",
  index = 1
)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Calculates correlation values between all the numeric pairs of variables in a dataset and puts it into
a dataframe. This dataframe can be used to visualise correlations.
}
\examples{
association_measure(iris,group_by = "Species",method = "distance")
}
