% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_measures.R
\name{association_measures}
\alias{association_measures}
\title{Calculates the measures of association (pearson, kendall, spearman, distance and mic)
between all variable pairs grouped by a grouping variable}
\usage{
association_measures(
  data = data,
  group_by = NULL,
  measures = c("pearson", "spearman", "kendall", "distance", "mic")
)
}
\arguments{
\item{data}{dataframe}

\item{group_by}{character vector specifying the grouping variable}

\item{measures}{character vector specifying measures of association}
}
\value{
dataframe
}
\description{
creates a dataframe with variable pairs and all the measures of association
}
