% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{calc_assoc_all}
\alias{calc_assoc_all}
\title{Calculates all association measures in the package for a dataset}
\usage{
calc_assoc_all(d, handle.na = T)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate measure of association}
}
\value{
tibble
}
\description{
Calculates all association measures for every variable pair in a dataset.
}
\examples{
calc_assoc_all(iris)
}
