% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{calc_assoc_by}
\alias{calc_assoc_by}
\title{Conditional Association measures for a dataset}
\usage{
calc_assoc_by(
  d,
  by = NULL,
  types = default_assoc(),
  handle.na = TRUE,
  include.overall = TRUE
)
}
\arguments{
\item{d}{dataframe}

\item{by}{a character string for the name of the conditional variable. Set to \emph{NULL} by default.}

\item{types}{a tibble for the measures to be calculated for different variable types. The default is
\emph{default_assoc()} which calculates Pearson's correlation if the variable pair is numeric,
Kendall's tau B if variable pair is ordered factor, canonical correlation if one is numeric and
other is a factor, and canonical correlation for any other variable pair.}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate measure of association.}

\item{include.overall}{If TRUE calculates the overall measure of association for every pair of variable and
includes it in the result.}
}
\value{
tibble
}
\description{
Calculates a measure of association for every variable pair at different levels of a third (conditional)
variable in a dataset.
}
\examples{
calc_assoc_by(iris,by="Species")
}
