% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dis.R
\name{cor_dis}
\alias{cor_dis}
\title{Find distance correlation data structure}
\usage{
cor_dis(data, index = index)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Calculates correlation values between all the numeric pairs of variables in a dataset and puts it into
a dataframe. This dataframe can be used to visualise correlations.
}
\examples{
cor_dis(iris[,-5])
}
