% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{default_assoc}
\alias{default_assoc}
\title{Default association measures calculated using \code{\link{calc_assoc}}}
\usage{
default_assoc()
}
\value{
tibble default association measures for pairs of variables
}
\description{
Gives a tibble for association measures for different types of variable pairs in a dataset.
The variable pairs (ordinal,factor) and (ordinal,numeric) are considered as factor pair and mixed pair
respectively. Used by \code{\link{calc_assoc}} for its \code{types} argument.
}
\examples{
default_assoc()
spearman_assoc <- default_assoc()
spearman_assoc$argList[[1]] <- list(method="spearman")
spearman_assoc
}
