% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_measure.R
\name{format_measure}
\alias{format_measure}
\title{Converts a measure (pearson correlation) matrix into a dataframe with the unique possible variable pairs
and value of the measure.}
\usage{
format_measure(measure_matrix)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Outputs a dataframe with variable pairs and measure of association.
This dataframe can be used to visualise the interesting measure of association.
}
\examples{
format_measure(measure_matrix)
}
