% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_plot.R
\name{funnel_plot}
\alias{funnel_plot}
\title{Display a linear layout plot for correlations between a designated response and the independent variables
grouped by a grouping variable}
\usage{
funnel_plot(measure_df, response)
}
\arguments{
\item{measure_df}{dataframe}

\item{response}{character vector as response}
}
\value{
plot
}
\description{
creates a funnel-like structure for correlations with a designated response
}
\examples{
funnel_plot(measure_df,response)
}
