% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_measures_Viz.R
\name{heatmap_measures_Viz}
\alias{heatmap_measures_Viz}
\title{Display a linear layout plot for the measures of association
between all the pairs of variables}
\usage{
heatmap_measures_Viz(association_measures, group_by = NULL)
}
\arguments{
\item{association_measures}{dataframe}
}
\value{
plot
}
\description{
creates a linear plot of all the variable pairs and measures of association in the form of a heatmap
}
\examples{
heatmap_mesaures_Viz(association_measures)

}
