% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\docType{data}
\name{measures}
\alias{measures}
\title{Association measure functions available in the package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 7 columns.
}
\usage{
measures
}
\value{
tibble
}
\description{
A tibble of association measure functions along with the types of variable pairs these functions
can be applied to. It also contains information regarding the packages used to calculate association
measures and the range of the measures calculated.
}
\keyword{datasets}
