% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_assoc.R
\name{order_assoc}
\alias{order_assoc}
\title{Ordering variables}
\usage{
order_assoc(assoc, group_var = group_var)
}
\arguments{
\item{assoc}{A tibble with association measures for different variable pairs in the dataset.}

\item{group_var}{a character vector for grouping variable. One of NULL, "by" or "measure_type"}
}
\value{
character vector representing the ordering of the variables
}
\description{
Calculates an ordering for the variables in a dataset.
}
\examples{
order_assoc(calc_assoc(iris))
order_assoc(calc_assoc(iris,"Species"))
order_assoc(calc_assoc(iris,"Species"),method="max_diff")
}
