% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{pairwise_1d_compare}
\alias{pairwise_1d_compare}
\title{Pairwise linear plot for comparing multiple measures of association}
\usage{
pairwise_1d_compare(
  df,
  measures = c("pearson", "spearman", "kendall", "cancor", "nmi"),
  var_order = "max_diff",
  limits = c(-1, 1),
  group_var = NULL
)
}
\arguments{
\item{df}{A tibble with the calculated multiple association measures for every variable pair in the dataset.}

\item{measures}{association measures to be calculated}

\item{var_order}{a character string for the ordering of the variables. Either  "max_diff"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}

\item{group_var}{a character string for the grouping variable. One of "by" (default) or "measure_type".}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a linear layout.
}
\examples{
pairwise_1d_compare(iris)
}
