% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{pairwise_1d_plot}
\alias{pairwise_1d_plot}
\title{Pairwise plot in a linear layout}
\usage{
pairwise_1d_plot(
  assoc,
  group_var = "by",
  fill = "default",
  var_order = "default",
  limits = c(-1, 1)
)
}
\arguments{
\item{assoc}{A tibble with the calculated association measures for every variable pair in the dataset..}

\item{group_var}{a character string for the grouping variable. One of "by" (default) or "measure_type".}

\item{fill}{a character string specifying the fill for the bars in the matrix plot. One of "default" (default)
for using levels of conditioning variable, "measure" for displaying a gradient or a color.}

\item{var_order}{a character string for the ordering of the variables. Either "default" (default) or "max_diff"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a linear layout.
}
\examples{
pairwise_1d_plot(calc_assoc_by(iris,"Species"))
}
