% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{pairwise_2d_plot}
\alias{pairwise_2d_plot}
\title{Pairs plot with association measures}
\usage{
pairwise_2d_plot(
  lassoc,
  uassoc = NULL,
  group_var = "by",
  fill = "default",
  var_order = "default",
  limits = c(-1, 1),
  sp = TRUE
)
}
\arguments{
\item{lassoc}{A tibble with the calculated association measures for the lower triangle of the matrix plot.}

\item{uassoc}{A tibble with the calculated association measures for the upper triangle of the matrix plot.
If \emph{NULL} (default) the matrix plot is symmetric.}

\item{group_var}{a character string for the grouping variable. One of "by" (default) or "measure_type".}

\item{fill}{a character string specifying the fill for the bars in the matrix plot. One of "default" (default)
for using levels of conditioning variable, "measure" for displaying a gradient or a color.}

\item{var_order}{a character string for the ordering of the variables. Either "default" (default) or "max_diff"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}

\item{sp}{a logical vector for highlighting instances of Simpson's paradox}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a matrix layout.
}
\examples{
pairwise_2d_plot(calc_assoc(iris))
pairwise_2d_plot(calc_assoc_by(iris,"Species"))
pairwise_2d_plot(calc_assoc_by(iris,"Species"),fill="measure_type")
}
