% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{pairwise_measures_compare}
\alias{pairwise_measures_compare}
\title{Pairwise linear plot for comparing multiple measures of association}
\usage{
pairwise_measures_compare(
  assoc,
  var_order = "max_diff",
  limits = c(-1, 1),
  group_var = NULL
)
}
\arguments{
\item{assoc}{A tibble with the calculated multiple association measures for every variable pair in the dataset..}

\item{var_order}{a character string for the ordering of the variables. Either  "max_diff"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}

\item{group_var}{a character string for the grouping variable. One of "by" (default) or "measure_type".}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a linear layout.
}
\examples{
updated_assoc <- update_assoc(num_pair = "tbl_cor",mixed_pair = "tbl_cancor",other_pair = "tbl_nmi")
a <- calc_assoc(iris,updated_assoc)
updated_assoc <- update_assoc(num_pair = "tbl_cor",num_pair_method = "spearman",mixed_pair = "tbl_nmi",
other_pair = "tbl_nmi")
b <- calc_assoc(iris,updated_assoc)
updated_assoc <- update_assoc(num_pair = "tbl_cor",num_pair_method = "kendall",mixed_pair = "tbl_nmi",
other_pair = "tbl_nmi")
c <- calc_assoc(iris,updated_assoc)
assoc <- unique(rbind(a,b,c))
pairwise_measures_compare(assoc)
}
