% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{plot_assoc_linear}
\alias{plot_assoc_linear}
\title{Pairwise plot in a linear layout}
\usage{
plot_assoc_linear(
  assoc,
  var_order = "default",
  plot_type = c("heatmap", "dotplot"),
  limits = c(-1, 1)
)
}
\arguments{
\item{assoc}{A tibble with the calculated association measures for every variable pair in the dataset..}

\item{var_order}{a character string for the ordering of the variables. Either "default" (default) or "max_diff"}

\item{plot_type}{a character string for specifying the plot type. One of "hearmap" or "dotplot"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a linear layout.
}
\examples{
plot_assoc_linear(calc_assoc(iris))
plot_assoc_linear(calc_assoc(iris,"Species"))
plot_assoc_linear(calc_assoc_all(iris))
}
