% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{plot_assoc_matrix}
\alias{plot_assoc_matrix}
\title{Pairwise plot in a matrix layout}
\usage{
plot_assoc_matrix(
  lassoc,
  uassoc = NULL,
  glyph = c("square", "circle"),
  var_order = "default",
  limits = c(-1, 1)
)
}
\arguments{
\item{lassoc}{A tibble with the calculated association measures for the lower triangle of the matrix plot.
Must be of class \code{pairwise}, \code{cond_pairwise} or \code{multi_pairwise}.}

\item{uassoc}{A tibble with the calculated association measures for the upper triangle of the matrix plot.
Must be of class \code{pairwise}, \code{cond_pairwise} or \code{multi_pairwise}.  If \emph{NULL} (default) the matrix plot is symmetric.}

\item{glyph}{A character string for the glyph to be used for lassoc with "pairwise" class. Either "square" or "circle".}

\item{var_order}{A character string for the variable order. Either "default" for ordering
using Dendser or a user provided variable order.}

\item{limits}{a numeric vector of length specifying the limits of the scale. Default is c(-1,1)}
}
\value{
A static \code{ggplot2} plot
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a matrix layout.
}
\examples{
plot_assoc_matrix(calc_assoc(iris))
plot_assoc_matrix(calc_assoc(iris,"Species"))
plot_assoc_matrix(calc_assoc_all(iris))
}
