% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_assoc.R
\name{sym_assoc}
\alias{sym_assoc}
\title{A tibble structure for a symmetric association matrix}
\usage{
sym_assoc(assoc)
}
\arguments{
\item{assoc}{A tibble or dataframe with calculated association measures for variable pairs.}
}
\value{
tibble
}
\description{
Creates a tibble with duplicated entries for each variable pair so that it can be used to create
a symmetric association matrix.
}
\examples{
sym_assoc(tbl_cor(iris))
}
