% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_ace}
\alias{tbl_ace}
\title{Pearson's Contingency Coefficient}
\usage{
tbl_ace(d, handle.na = T, ...)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates Pearson's Contingency coefficient for every variable pair in a dataset.
}
\examples{
tbl_chi(iris)
}
