% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_cancor}
\alias{tbl_cancor}
\title{Canonical correlation}
\usage{
tbl_cancor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble with canonical correlation for every variable pair
}
\description{
Calculates canonical correlation for every variable pair in a dataset.
}
\examples{
tbl_cancor(iris)
}
