% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_chi}
\alias{tbl_chi}
\title{Pearson's Contingency Coefficient}
\usage{
tbl_chi(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates Pearson's Contingency coefficient for every variable pair in a dataset.
}
\examples{
tbl_chi(iris)
}
