% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_cor}
\alias{tbl_cor}
\title{Pearson, Spearman or Kendall correlation}
\usage{
tbl_cor(d, method = "pearson", handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{method}{A character string for the correlation coefficient to be calculated. Either "pearson" (default),
"spearman", or "kendall"}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble with calculated association measure for every numeric variable pair
}
\description{
Calculates one of either pearson, spearman or kendall correlation for every numeric variable pair in a dataset.
}
\examples{
tbl_cor(iris)
tbl_cor(iris, method="kendall")
tbl_cor(iris, method="spearman")
}
