% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_easy}
\alias{tbl_easy}
\title{Easystats correlations}
\usage{
tbl_easy(d, method = "pearson", handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{method}{a character string for the correlation coefficient to be calculated. One of "pearson" (default),
"spearman", "kendall", "biserial", "polychoric", "tetrachoric", "biweight", "distance",
"percentage" (for percentage bend correlation), "blomqvist" (for Blomqvist's coefficient),
"hoeffding" (for Hoeffding's D), "gamma", "gaussian" (for Gaussian Rank correlation) or
"shepherd" (for Shepherd's Pi correlation). Setting "auto" will select the most most relevant
method depending on the variable types in the dataset.}

\item{handle.na}{NA handling not available}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates one of the many correlation coefficients available with easystats package
for variable pairs in a dataset.
}
\examples{
tbl_easy(iris)
tbl_easy(iris,method="hoeffding")
}
