% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_gkGamma}
\alias{tbl_gkGamma}
\title{Goodman Kruskal's Gamma}
\usage{
tbl_gkGamma(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
A tibble with ordinal variable pairs and Goodman Kruskal's Gamma coefficient
}
\description{
Calculates Goodman Kruskal's Gamma coefficient for every ordinal variable pair in a dataset.
}
\details{
The Goodman Kruskal's Gamma coefficient is calculated using \code{\link[DescTools]{GoodmanKruskalGamma}}
function from the \code{DescTools} package.
}
\examples{
tbl_gkGamma(iris)
}
