% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_gkTau}
\alias{tbl_gkTau}
\title{Goodman Kruskal's Tau}
\usage{
tbl_gkTau(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates Goodman Kruskal's Tau coefficient for every variable pair in a dataset.
}
\examples{
tbl_gkTau(iris)
}
