% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_mine}
\alias{tbl_mine}
\title{MINE family measures}
\usage{
tbl_mine(d, method = "mic", handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{method}{character string for the MINE measure to be calculated. Either "mic" (default), "mas", "mev",
"mcn", or "mic-r2"}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates MINE family measures for every numeric variable pair in a dataset.
}
\examples{
tbl_mine(iris)
tbl_mine(iris, method="mas")
}
