% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_nmi.R
\name{tbl_nmi}
\alias{tbl_nmi}
\title{Normalized mutual information}
\usage{
tbl_nmi(d, handle.na = T, ...)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{...}{in progress}

\item{handle.NA}{If TRUE uses pairwise complete observations to calculate correlation coefficient}
}
\value{
tibble
}
\description{
Calculates normalized mutual information for every variable pair in a dataset.
}
\examples{
tbl_nmicor(iris)
}
