% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures_funs.R
\name{tbl_passo}
\alias{tbl_passo}
\title{Patial Association}
\usage{
tbl_passo(d, method = "pearson", handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{method}{a character string for the correlation coefficient to be calculated. Either "pearson" (default),
"spearman", or "kendall"}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{in progress}
}
\value{
tibble
}
\description{
Calculates partial association coefficient for every ordinal variable pair in a dataset.
}
