% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_polycor}
\alias{tbl_polycor}
\title{Polychoric correlation}
\usage{
tbl_polycor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates Polychoric correlation for every variable pair in a dataset.
}
\examples{
tbl_polycor(iris)
}
