% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_scag}
\alias{tbl_scag}
\title{Graph-theoretic scagnostics measures}
\usage{
tbl_scag(d, scagnostic = "Outlying", handle.na = T, ...)
}
\arguments{
\item{d}{A dataframe}

\item{scagnostic}{a character string for the scagnostic to be calculated. One of "Outlying",
"Stringy", "Striated", "Clumpy", "Sparse", "Skewed", "Convex", "Skinny" or "Monotonic"}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
A tibble with one of the nine scagnostic measures for every numeric variable pair
}
\description{
Calculates scagnostic measure for every numeric variable pair in a dataset.
}
\details{
The scagnostic measures are calculated using \code{\link[scagnostics]{scagnostics}}
function from the \code{scagnostics} package.
}
\examples{
tbl_scag(iris)
}
\references{
Wilkinson, Leland, Anushka Anand, and Robert Grossman.
"Graph-theoretic scagnostics."
Information Visualization, IEEE Symposium on. IEEE Computer Society, 2005
}
