% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_uncertainty}
\alias{tbl_uncertainty}
\title{Uncertainty coefficient}
\usage{
tbl_uncertainty(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble with every nominal variable pair and uncertainty coefficient value.
}
\description{
Calculates uncertainty coefficient for every nominal variable pair in a dataset.
}
\details{
The Uncertainty coefficient is calculated using \code{UncertCoef} function from the
\code{DescTools} package.
}
\examples{
tbl_uncertainty(iris)
}
