% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{update_assoc}
\alias{update_assoc}
\title{A user friendly function for changing association measures}
\usage{
update_assoc(
  num_pair = "tbl_cor",
  num_pair_method = NULL,
  ordered_pair = "tbl_tau",
  ordered_pair_method = NULL,
  mixed_pair = "tbl_cancor",
  mixed_pair_method = NULL,
  other_pair = "tbl_cancor",
  other_pair_method = NULL,
  ...
)
}
\arguments{
\item{num_pair}{a measure(s) function for numeric pair of variables}

\item{num_pair_method}{a character string specifying the measure to be calculated using num_pair}

\item{ordered_pair}{a measure(s) function for ordered pair of variables}

\item{ordered_pair_method}{a character string specifying the measure to be calculated using ordered_pair}

\item{mixed_pair}{a measure(s) function for mixed pair of variables}

\item{mixed_pair_method}{a character string specifying the measure to be calculated using mixed_pair}

\item{other_pair}{a measure(s) function for other pair of variables}

\item{other_pair_method}{a character string specifying the measure to be calculated using other_pair}

\item{...}{in progress}
}
\value{
tibble
}
\description{
Creates a tibble for different measures of association for different variable types of a dataset.
}
\examples{
updated_assoc <- update_assoc(num_pair="tbl_cor", num_pair_method=NULL,
ordered_pair="tbl_tau", ordered_pair_method=NULL,
mixed_pair="tbl_cancor", mixed_pair_method=NULL,
other_pair="tbl_cancor", other_pair_method=NULL)
calc_assoc(iris,updated_assoc)
}
