% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{write_own_assoc}
\alias{write_own_assoc}
\title{A user friendly function for changing association measures}
\usage{
write_own_assoc(assoc_fun, fun_index, argList = NULL)
}
\arguments{
\item{assoc_fun}{the association function user would like to use}

\item{fun_index}{an index for the function representing the row on which function needs to be updated}

\item{argList}{a character string specifying the method to use when multiple methods present in assoc_fun}
}
\value{
tibble
}
\description{
Creates a tibble for different measures of association for different variable types of a dataset.
}
