% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{assoc_tibble.matrix}
\alias{assoc_tibble.matrix}
\title{A tibble structure for a measure of association}
\usage{
\method{assoc_tibble}{matrix}(data, measure_type = "?")
}
\arguments{
\item{data}{matrix A symmetric matrix of the variabels of a dataset with entries representing the measure of
association}

\item{measure_type}{a character string indicating the measure of association}
}
\value{
tibble
}
\description{
Creates a tibble for every variable pair in a dataset with a measure of association
}
