% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_tibble.R
\name{assoc_tibble}
\alias{assoc_tibble}
\title{A tibble structure for a measure of association}
\usage{
assoc_tibble(d, measure_type = "?")
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables or a dataframe with calculated
measure of association for every variable pair}

\item{measure_type}{a character string indicating the measure of association}

\item{m}{matrix A symmetric matrix of the variabels of a dataset with entries representing the measure of
association}
}
\value{
tibble
}
\description{
Creates a tibble for every variable pair in a dataset with a measure of association
}
\examples{
assoc_tibble(cor(iris[,1:4]), measure_type="pearson")
assoc_tibble(iris)
}
