% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_display.R
\name{bar_display}
\alias{bar_display}
\title{Display a matrix layout plot for correlations between pairs of variables grouped by a grouping variable}
\usage{
bar_display(association_measure)
}
\arguments{
\item{data}{dataframe}
}
\value{
plot
}
\description{
creates a pairwise plot of correlations for all the numeric variables grouped by a grouping variable
}
\examples{
bar_display(iris,group_by = "Species",method = "distance")
}
