% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_assoc.R
\name{calc_assoc}
\alias{calc_assoc}
\title{Association measures for a dataset}
\usage{
calc_assoc(d, types = default_assoc(), handle.na = TRUE)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{types}{a tibble for the measures to be calculated for different variable types. The default is
\emph{default_assoc()} which calculates Pearson's correlation if the variable pair is numeric,
Kendall's tau B if variable pair is ordered factor, canonical correlation if one is numeric and
other is a factor, and canonical correlation for any other variable pair.}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate measure of association}
}
\value{
tibble
}
\description{
Calculates a measure of association for every variable pair in a dataset.
}
\examples{
cal_assoc(iris)
}
