% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_assoc.R
\name{default_assoc}
\alias{default_assoc}
\title{A user defined function for association measures}
\usage{
default_assoc()
}
\value{
tibble
}
\description{
Creates a tibble for different measures of association for different variable types of a dataset.
}
\examples{
default_assoc()
spearman_assoc <- default_assoc()
spearman_assoc$argList[[1]] <- list(method="spearman")
spearman_assoc()
}
