% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_structure.R
\name{matrix_assoc}
\alias{matrix_assoc}
\title{Symmetric association matrix}
\usage{
matrix_assoc(assoc, group = "overall")
}
\arguments{
\item{assoc}{A tibble or dataframe with calculated association measures for variable pairs.}

\item{group}{A character string specifying the level of the conditioning variable for
which matrix to be created}
}
\value{
tibble
}
\description{
Creates a symmetric association matrix of calcualted association measures among pairs of variables.
}
\examples{
matrix_assoc(tbl_cor(iris))
}
