% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_assoc.R
\name{matrix_assoc}
\alias{matrix_assoc}
\title{Symmetric association matrix}
\usage{
matrix_assoc(assoc)
}
\arguments{
\item{assoc}{A tibble or dataframe with calculated association measures for variable pairs.}
}
\value{
tibble
}
\description{
Creates a symmetric association matrix of calcualted association measures among pairs of variables.
}
\examples{
matrix_assoc(tbl_cor(iris))
}
