% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\docType{data}
\name{measures}
\alias{measures}
\title{Association functions available in the package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 7 columns.
}
\usage{
measures
}
\value{
tibble
}
\description{
List of all the functions available to calculate association measures in the package.
}
\keyword{datasets}
