% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_assoc.R
\name{order_assoc}
\alias{order_assoc}
\title{Ordering variables}
\usage{
order_assoc(assoc, method = "default", group_var = NULL)
}
\arguments{
\item{assoc}{A tibble with association measures for different variable pairs in the dataset.}

\item{method}{a character string for the method to be used for ordering. One of "default" (default) which uses
average linkage clustering on the overall measure values or "max_diff" which uses average
linkage clustering on the maximum difference of the measure valuesfor variable pairs at different
levels of a conditioning variable.}

\item{group_var}{in progress}
}
\value{
character vector representing the ordering of the variables
}
\description{
Calculates an ordering for the variables in a dataset.
}
\examples{
order_assoc(calc_assoc(iris))
order_assoc(calc_assoc_by(iris,"Species"))
order_assoc(calc_assoc_by(iris,"Species"),method="max_diff")
}
