% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{plot_assoc_linear}
\alias{plot_assoc_linear}
\title{Pairwise plot in a linear layout}
\usage{
plot_assoc_linear(
  assoc,
  pair_order = "default",
  plot_type = c("heatmap", "dotplot"),
  limits = c(-1, 1)
)
}
\arguments{
\item{assoc}{A tibble with the calculated association measures for every variable pair in the dataset.
Must be of class \code{pairwise}, \code{cond_pairwise} or \code{multi_pairwise}.}

\item{pair_order}{a character string for ordering of the pairs of variables in linear layout. One of "default" (default) or "max_diff".
When set to "default", pairs are arranged in decreasing order of the absolute value of measure or measures (when multiple measure per pair are present).
When set to "max_diff", the ordering is only applicable to assoc with class cond_pairwise or multi_pairwise and the pairs are ordered in descending order of the maximum difference calculated among the measures.}

\item{plot_type}{a character string for specifying the type of plot an analyst wants. One of "heatmap" or "dotplot".}

\item{limits}{a numeric vector specifying the limits of the scale. Default is c(-1,1)}
}
\value{
A static \code{ggplot2} plot
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a linear layout.
}
\examples{
plot_assoc_linear(calc_assoc(iris))
plot_assoc_linear(calc_assoc(iris,"Species"))
plot_assoc_linear(calc_assoc_all(iris))
}
