% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plot.R
\name{plot_assoc_matrix}
\alias{plot_assoc_matrix}
\title{Pairs plot with in a matrix layout}
\usage{
plot_assoc_matrix(
  lassoc,
  uassoc = NULL,
  glyph = c("square", "circle"),
  group_var = NULL,
  fill = "default",
  var_order = "default",
  limits = c(-1, 1)
)
}
\arguments{
\item{lassoc}{A tibble with the calculated association measures for the lower triangle of the matrix plot.}

\item{uassoc}{A tibble with the calculated association measures for the upper triangle of the matrix plot.
If \emph{NULL} (default) the matrix plot is symmetric.}

\item{glyph}{A character string for the glyph to be used. Either "square" or "circle"}

\item{group_var}{a character string for the grouping variable. One of NULL (default), "by"  or "measure_type".}

\item{fill}{a character string specifying the fill for the bars in the matrix plot. One of "default" (default)
for using levels of conditioning variable, "measure" for displaying a gradient or a color.}

\item{var_order}{a character string for the ordering of the variables. Either "default" (default) or "max_diff"}

\item{limits}{a numeric vector of length $2$ specifying the limits of the scale. Default is c(-1,1)}
}
\description{
Plots the calculated measures of association among different variable pairs for a dataset in a matrix layout.
}
\examples{
plot_assoc_matrix(calc_assoc(iris))
plot_assoc_matrix(calc_assoc(iris,"Species"), group_var = "by")
plot_assoc_matrix(calc_assoc_all(iris),group_var ="measure_type")
}
