% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_cancor.R
\name{tbl_cancor}
\alias{tbl_cancor}
\title{Canonical correlation}
\usage{
tbl_cancor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{in progress}
}
\value{
tibble
}
\description{
Calculates canonical correlation for every variable pair in a dataset.
}
\examples{
tbl_cancor(iris)
}
