% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_chi}
\alias{tbl_chi}
\title{Pearson's Contingency Coefficient}
\usage{
tbl_chi(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
A tibble with calculated Pearson's contingency coefficient for every nominal variable
pair
}
\description{
Calculates Pearson's Contingency coefficient for every nominal variable pair in a dataset.
}
\details{
The Pearson's contingency coefficient is calculated using \code{\link[DescTools]{ContCoef}}
function from the \code{DescTools} package.
}
\examples{
tbl_chi(iris)
}
