% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_dcor}
\alias{tbl_dcor}
\title{Distance correlation}
\usage{
tbl_dcor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble with distance correlation for every numeric variable pair
}
\description{
Calculates distance correlation for every numeric variable pair in a dataset.
}
\examples{
tbl_dcor(iris)
}
