% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_easy.R
\name{tbl_easy}
\alias{tbl_easy}
\title{Easystats correlations}
\usage{
tbl_easy(d, method = "pearson", handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{method}{a character string for the correlation coefficient to be calculated. One of "pearson" (default),
"spearman", "kendall", "biserial", "polychoric", "tetrachoric", "biweight", "distance",
"percentage" (for percentage bend correlation), "blomqvist" (for Blomqvist's coefficient),
"hoeffding" (for Hoeffding's D), "gamma", "gaussian" (for Gaussian Rank correlation) or
"shepherd" (for Shepherd's Pi correlation). Setting "auto" will select the most most relevant
method depending on the variable types in the dataset.}

\item{handle.na}{NA handling not available}

\item{...}{in progress}
}
\value{
tibble
}
\description{
Calculates one of the many correlation coefficients available with easystats package
for variable pairs in a dataset.
}
\examples{
tbl_easy(iris)
tbl_easy(iris,method="hoeffding")
}
