% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_gkGamma}
\alias{tbl_gkGamma}
\title{Goodman Kruskal's Gamma}
\usage{
tbl_gkGamma(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates Goodman Kruskal's Gamma coefficient for every variable pair in a dataset.
}
\examples{
tbl_gkGamma(iris)
}
