% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_gkTau}
\alias{tbl_gkTau}
\title{Goodman Kruskal's Tau}
\usage{
tbl_gkTau(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
A tibble with Goodman Kruskal's Tau for every nominal variable pair
}
\description{
Calculates Goodman Kruskal's Tau coefficient for every ordinal variable pair in a dataset.
}
\details{
The Goodman Kruskal's Tau coefficient is calculated using \code{\link[DescTools]{GoodmanKruskalTau}}
function from the \code{DescTools} package.
}
\examples{
tbl_gkTau(iris)
}
