% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_mine}
\alias{tbl_mine}
\title{MINE family measures}
\usage{
tbl_mine(d, method = "mic", handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe or tibble}

\item{method}{character string for the MINE measure to be calculated. Either "mic" (default), "mas", "mev",
"mcn", or "mic-r2"}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble
}
\description{
Calculates MINE family measures for every numeric variable pair in a dataset.
}
\details{
The measures are calculated using \code{\link[minerva]{mine}} from \code{minerva}
}
\examples{
tbl_mine(iris)
tbl_mine(iris, method="mas")
}
\references{
Reshef, David N., et al. "Detecting novel associations in large data sets."
science 334.6062 (2011): 1518-1524
}
