% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_nmi}
\alias{tbl_nmi}
\title{Normalized mutual information}
\usage{
tbl_nmi(d, handle.na = T, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate normalized mutual information}

\item{...}{other arguments}
}
\value{
A tibble
}
\description{
Calculates normalized mutual information for every variable pair in a dataset.
}
\details{
The normalized mutual information is calculated using \code{\link[DescTools]{MutInf}} and  \code{\link[DescTools]{Entropy}} from DescTools
}
\examples{
tbl_nmi(iris)
}
