% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_polycor}
\alias{tbl_polycor}
\title{Polychoric correlation}
\usage{
tbl_polycor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe or tibble}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient}

\item{...}{other arguments}
}
\value{
A tibble with polychoric correlation for ordinal variable pairs
}
\description{
Calculates Polychoric correlation using from polycor for every ordinal variable pair in a dataset.
}
\details{
The polychoric correlation is calculated using the \code{\link[polycor]{polychor}} function from the
\code{polycor} package
}
\examples{
tbl_polycor(iris)
}
