% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_scag}
\alias{tbl_scag}
\title{Graph-theoretic scagnostics measures}
\usage{
tbl_scag(d, scagnostic = "Outlying", handle.na = T, ...)
}
\arguments{
\item{d}{dataframe}

\item{scagnostic}{a character string for the scagnostic to be calculated. One of "Outlying",
"Stringy", "Striated", "Clumpy", "Sparse", "Skewed", "Convex", "Skinny" or "Monotonic"}

\item{handle.na}{If TRUE uses pairwise complete observations.}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Calculates scagnostic measure for every variable pair in a dataset.
}
\examples{
tbl_scag(iris)
}
