% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_measures.R
\name{tbl_tau}
\alias{tbl_tau}
\title{Kendall's tau A, B, C and Kendall's W}
\usage{
tbl_tau(d, method = c("B", "A", "C", "W"), ...)
}
\arguments{
\item{d}{A dataframe}

\item{method}{A character string for the correlation coefficient to be calculated. Either "B" (default),
"A", "C" or "W"}

\item{...}{other arguments}
}
\value{
A tibble with ordinal variable pairs along with one of either Kendall's tau A, B, C or
Kendall's W measure
}
\description{
Calculates one of either Kendall's tau A, B, C or Kendall's W for every ordinal variable pair in a dataset.
}
\details{
The association measures Kendall's tau A, B, C or Kendall's W are calculated using \code{\link[DescTools]{KendallTauA}},
\code{\link[DescTools]{KendallTauB}}, \code{\link[DescTools]{StuartTauC}} or \code{\link[DescTools]{KendallW}} respectively,from the
\code{DescTools} package.
}
\examples{
tbl_tau(iris)
tbl_tau(iris, method="A")
tbl_tau(iris, method="C")
tbl_tau(iris, method="W")
}
