% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_uncertainty.R
\name{tbl_uncertainty}
\alias{tbl_uncertainty}
\title{Uncertainty coefficient}
\usage{
tbl_uncertainty(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{dataframe A dataset for exploring association among the variables.}

\item{...}{in progress}

\item{handle.NA}{If TRUE uses pairwise complete observations to calculate correlation coefficient}
}
\value{
tibble
}
\description{
Calculates uncertainty coefficient for every variable pair in a dataset.
}
\examples{
tbl_uncertainty(iris)
}
