% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_association.R
\name{update_assoc}
\alias{update_assoc}
\title{A user friendly function for changing association measures}
\usage{
update_assoc(
  default = default_assoc(),
  num_pair = NULL,
  num_pair_argList = NULL,
  factor_pair = NULL,
  factor_pair_argList = NULL,
  ordered_pair = NULL,
  ordered_pair_argList = NULL,
  mixed_pair = NULL,
  mixed_pair_argList = NULL,
  ...
)
}
\arguments{
\item{default}{default measure functions for different variable pairs. set to \code{\link[=default_assoc]{default_assoc()}}}

\item{num_pair}{a measure(s) function for numeric pair of variables, default is NULL}

\item{num_pair_argList}{a character string specifying the measure to be calculated using num_pair, default is NULL}

\item{factor_pair}{a measure(s) function for factor pair of variables, default is NULL}

\item{factor_pair_argList}{a character string specifying the measure to be calculated using factor_pair, default is NULL}

\item{ordered_pair}{a measure(s) function for ordered pair of variables, default is NULL}

\item{ordered_pair_argList}{a character string specifying the measure to be calculated using ordered_pair, default is NULL}

\item{mixed_pair}{a measure(s) function for mixed pair of variables, default is NULL}

\item{mixed_pair_argList}{a character string specifying the measure to be calculated using mixed_pair, default is NULL}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
Creates a tibble for different measures of association for different variable types of a dataset.
}
\examples{
updated_assoc <- update_assoc(num_pair="tbl_cor", num_pair_argList="spearman",
ordered_pair="tbl_tau",mixed_pair="tbl_nmi",factor_pair="tbl_cancor")
calc_assoc(iris,types=updated_assoc)
}
